Move RocksDB stats code to a separate private module.
